+++ database/seeders/ProjectSeeder.php
@@ -2,7 +2,6 @@
 
 namespace Database\Seeders;
 
-use DB;
 use App\Models\Project;
 use Illuminate\Database\Seeder;
 
@@ -13,10 +12,7 @@ class ProjectSeeder extends Seeder
      */
     public function run(): void
     {
-        DB::statement('SET FOREIGN_KEY_CHECKS=0;');
         Project::truncate();
-        DB::statement('SET FOREIGN_KEY_CHECKS=1;');
-
         Project::factory(5)->create();
     }
 }
